+++ lib/a1.dart
@@ -26,5 +26,5 @@ library;
 export 'src/a1.dart';
 export 'src/a1_partial.dart';
 export 'src/a1_range.dart';
-export 'src/a1_range_binary_search.dart';
+export 'src/a1_range_search.dart';
 export 'src/a1_reference.dart';