+++ lib/src/a1_partial.dart
@@ -1,3 +1,4 @@
+// ignore_for_file: public_member_api_docs, sort_constructors_first
 // Copyright (c) 2024, Scott Horn.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
@@ -7,6 +8,7 @@ import 'package:a1/a1.dart';
 class A1Partial implements Comparable {
   /// Empty/all [A1Partial]
   static A1Partial all = A1Partial(null, null);
+  static final _maxInt = -1 >>> 1;
 
   /// letters of the A1 or null for all columns
   final String? letters;
@@ -68,6 +70,56 @@ class A1Partial implements Comparable {
     return thisA1.compareTo(otherA1);
   }
 
+  /// Conveniance function for determining range from a list of partials
+  static A1Range rangefromList(
+      List<A1Partial?> listFrom, List<A1Partial?> listTo,
+      {Object? tag, A1? anchor}) {
+    const maxInt = -1 >>> 1;
+
+    // clean the nulls and create from list
+    final fromItems = listFrom.where((element) => element != null);
+    final fromColumns =
+        fromItems.map((A1Partial? e) => e?.column ?? 0).toList();
+    fromColumns.sort();
+    final fromColumn = fromColumns.first;
+    final fromRows = fromItems.map((A1Partial? e) => e?.row ?? 0).toList();
+    fromRows.sort();
+    final fromRow = fromRows.first;
+
+    // clean the nulls and create to list
+    final toItems = listTo.where((element) => element != null);
+    final toColumns =
+        toItems.map((A1Partial? e) => e?.column ?? maxInt).toList();
+    toColumns.sort();
+
+    final toColumn = toColumns.last;
+    final toRows = toItems.map((A1Partial? e) => e?.row ?? maxInt).toList();
+    toRows.sort();
+    final toRow = toRows.last;
+
+    return A1Range.fromPartials(
+      A1Partial.fromVector(fromColumn, fromRow),
+      A1Partial.fromVector(
+          toColumn == maxInt ? null : toColumn, toRow == maxInt ? null : toRow),
+      tag: tag,
+      anchor: anchor,
+    );
+  }
+
+  /// Conveniance function for determining minimum of a list of partials
+  static A1Partial min(Iterable<A1Partial?> list) {
+    final items = list.where((element) => element != null).toList();
+    items.sort();
+    return items.first!;
+  }
+
+  /// Conveniance function for determining maximum of a list of partials
+  static A1Partial max(Iterable<A1Partial?> list) {
+    final items = list.where((element) => element != null).toList();
+    items.sort();
+    return items.last!;
+  }
+
   /// Return the column as a zero based [int] or null
   int? get column {
     if (letters == null) return null;
@@ -116,4 +168,39 @@ class A1Partial implements Comparable {
 
   /// if this partial represents either a whole row or column
   bool get isWholeRowOrColumn => isWholeColumn || isWholeRow;
+
+  /// Utility for copying
+  A1Partial vectorCopyWith({
+    int? column,
+    int? row,
+  }) {
+    return A1Partial.fromVector(
+      column ?? this.column,
+      row ?? this.row,
+    );
+  }
+
+  /// utility for moving right
+  A1Partial get right => vectorCopyWith(
+      column: column != null && column! < _maxInt ? column! + 1 : null);
+
+  /// utility for moving left
+  A1Partial get left => vectorCopyWith(
+      column: column == null
+          ? null
+          : column! > 0
+              ? column! - 1
+              : 0);
+
+  /// utility for moving down
+  A1Partial get down =>
+      vectorCopyWith(row: row != null && row! < _maxInt ? row! + 1 : null);
+
+  /// utility for moving up
+  A1Partial get up => vectorCopyWith(
+      row: row == null
+          ? null
+          : row! > 0
+              ? row! - 1
+              : 0);
 }